+++ src/cuddcppbman.cc
@@ -643,7 +643,7 @@ CuddCPPBMan::visualize(BooleRepr * abr)
 {
     CuddCPPBooleRepr * a = dcast(abr);
     char buffer[200];
-    sprintf(buffer, "cudd_bdd%u.dot", (unsigned int) abr);
+    sprintf(buffer, "cudd_bdd%lu.dot", (size_t) abr);
     FILE *outfile;
     outfile = fopen(buffer, "w");
     std::vector<BDDVar> v;
